pyproject.toml
[tool.poetry]
name = "cv"
version = "0.1.0"
description = ""
authors = ["emd1tey <emditey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
anyio = "4.4.0"
Babel = "2.15.0"
black = "24.3.0"
bracex = "2.4"
CacheControl = "0.14.0"
certifi = "2024.6.2"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
cyclonedx-python-lib = "7.4.1"
defusedxml = "0.7.1"
distlib = "0.3.8"
dnspython = "2.6.1"
email-validator = "2.1.1"
fastapi = "0.111.0"
fastapi-cli = "0.0.4"
filelock = "3.14.0"
fonttools = "4.53.0"
fpdf2 = "2.7.9"
geographiclib = "2.0"
geopy = "2.4.1"
ghp-import = "2.1.0"
h11 = "0.14.0"
html5lib = "1.1"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
identify = "2.5.36"
idna = "3.7"
isort = "5.13.2"
Jinja2 = "3.1.4"
license-expression = "30.3.0"
Markdown = "3.6"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
mdurl = "0.1.2"
mergedeep = "1.3.4"
mkdocs = "1.6.0"
mkdocs-awesome-pages-plugin = "2.9.2"
mkdocs-get-deps = "0.2.0"
mkdocs-material = "9.5.26"
mkdocs-material-extensions = "1.3.1"
msgpack = "1.0.8"
mypy-extensions = "1.0.0"
natsort = "8.4.0"
nodeenv = "1.9.1"
orjson = "3.10.3"
packageurl-python = "0.15.1"
packaging = "24.0"
paginate = "0.5.6"
pathspec = "0.12.1"
pillow = "10.3.0"
pip-api = "0.0.33"
pip-requirements-parser = "32.0.1"
pip-audit = "2.7.3"
platformdirs = "4.2.2"
pre-commit = "3.7.1"
py-serializable = "1.0.3"
pydantic = "2.7.3"
pydantic-core = "2.18.4"
Pygments = "2.18.0"
pymdown-extensions = "10.8.1"
pyparsing = "3.1.2"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
PyYAML = "6.0.1"
pyyaml-env-tag = "0.1"
regex = "2024.5.15"
requests = "2.32.3"
rich = "13.7.1"
ruff = "0.4.9"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
starlette = "0.37.2"
toml = "0.10.2"
typer = "0.12.3"
typing-extensions = "4.12.2"
ujson = "5.10.0"
urllib3 = "2.2.1"
uvicorn = "0.30.1"
uvloop = "0.19.0"
virtualenv = "20.26.2"
watchdog = "4.0.1"
watchfiles = "0.22.0"
wcmatch = "8.5.2"
webencodings = "0.5.1"
websockets = "12.0"
annotated-types = "0.7.0"
aiofiles = "^23.2.1"
prometheus-client = "^0.20.0"
opentelemetry-distro = "^0.46b0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-instrumentation-logging = "^0.46b0"
opentelemetry-exporter-otlp = "^1.25.0"
elastic-apm = "^6.22.3"
opentelemetry-exporter-otlp-proto-grpc = "^1.25.0"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-proto = "^1.25.0"
opentelemetry-exporter-prometheus = "^0.46b0"
gunicorn = "^22.0.0"
opentelemetry-instrumentation-requests = "^0.46b0"
opentelemetry-exporter-gcp-trace = "^1.6.0"
elasticsearch = {extras = ["async"], version = "^8.14.0"}
pytest = "^8.2.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.ruff]
line-length = 88
